Added GraalVM lab
