+++ labs/unicorn-location-api/final/unicorn-location-api-final/UnicornLocationFunction/pom.xml
@@ -15,7 +15,7 @@
         <dependency>
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>dynamodb</artifactId>
-            <version>2.17.214</version>
+            <version>2.17.282</version>
             <exclusions>
                 <exclusion>
                     <groupId>software.amazon.awssdk</groupId>
@@ -32,7 +32,7 @@
         <dependency>
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>aws-crt-client</artifactId>
-            <version>2.17.214-PREVIEW</version>
+            <version>2.17.282-PREVIEW</version>
         </dependency>
         <dependency>
             <groupId>com.amazonaws</groupId>
