+++ labs/unicorn-store/infrastructure/db-setup/pom.xml
@@ -8,7 +8,7 @@
     <name>AWS Lambda function to setup the Amazon RDS Instance</name>
 
     <properties>
-        <aws.java.sdk.version>2.17.281</aws.java.sdk.version>
+        <aws.java.sdk.version>2.17.282</aws.java.sdk.version>
     </properties>
 
     <dependencyManagement>
