+++ labs/unicorn-store/software/alternatives/unicorn-store-basic/pom.xml
@@ -14,7 +14,7 @@
 		<dependency>
 			<groupId>software.amazon.awssdk</groupId>
 			<artifactId>eventbridge</artifactId>
-			<version>2.17.281</version>
+			<version>2.17.282</version>
 			<exclusions>
 				<exclusion>
 					<groupId>software.amazon.awssdk</groupId>
@@ -29,7 +29,7 @@
 		<dependency>
 			<groupId>software.amazon.awssdk</groupId>
 			<artifactId>aws-crt-client</artifactId>
-			<version>2.17.261-PREVIEW</version>
+			<version>2.17.282-PREVIEW</version>
 		</dependency>
 		<dependency>
 			<groupId>org.postgresql</groupId>
