+++ labs/unicorn-store/software/alternatives/unicorn-store-micronaut/pom.xml
@@ -117,7 +117,7 @@
     <dependency>
       <groupId>software.amazon.awssdk</groupId>
       <artifactId>eventbridge</artifactId>
-      <version>2.17.281</version>
+      <version>2.17.282</version>
       <exclusions>
         <exclusion>
           <groupId>software.amazon.awssdk</groupId>
@@ -132,7 +132,7 @@
     <dependency>
       <groupId>software.amazon.awssdk</groupId>
       <artifactId>aws-crt-client</artifactId>
-      <version>2.17.261-PREVIEW</version>
+      <version>2.17.282-PREVIEW</version>
     </dependency>
   </dependencies>
