--- labs/unicorn-store/software/alternatives/unicorn-store-spring-final/src/main/java/com/unicorn/store/StoreApplication.java
@@ -1,25 +1,0 @@
-package com.unicorn.store;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
-
-@SpringBootApplication
-@EnableJdbcRepositories
-public class StoreApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(StoreApplication.class, args);
-	}
-
-	@Bean
-	public ObjectMapper getObjectMapper() {
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-		return objectMapper;
-	}
-
-}
