--- labs/unicorn-store/software/alternatives/unicorn-store-spring-final/src/main/java/com/unicorn/store/exceptions/PublisherException.java
@@ -1,7 +1,0 @@
-package com.unicorn.store.exceptions;
-
-public class PublisherException extends RuntimeException{
-    public PublisherException(String errorMessage, Throwable err) {
-        super(errorMessage, err);
-    }
-}
