--- labs/unicorn-store/software/alternatives/unicorn-store-spring-final/src/main/java/com/unicorn/store/exceptions/ResourceNotFoundException.java
@@ -1,5 +1,0 @@
-package com.unicorn.store.exceptions;
-
-public class ResourceNotFoundException extends RuntimeException{
-
-}
