--- labs/unicorn-store/software/alternatives/unicorn-store-spring-final/src/main/java/com/unicorn/store/model/UnicornEventType.java
@@ -1,5 +1,0 @@
-package com.unicorn.store.model;
-
-public enum UnicornEventType {
-    UNICORN_CREATED, UNICORN_UPDATED, UNICORN_DELETED
-}
