--- labs/unicorn-store/software/alternatives/unicorn-store-spring-final/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-#spring.datasource.url=<UnicornStoreSpringStack will set it via the env variable SPRING_DATASOURCE_URL at deployment time>
-spring.datasource.username=postgres
-#spring.datasource.password=<UnicornStoreSpringStack will set it via the env variable SPRING_DATASOURCE_PASSWORD at deployment time>
-server.error.include-message=always
-spring.jpa.open-in-view=false
-spring.datasource.hikari.maximumPoolSize=1
-spring.datasource.hikari.data-source-properties.preparedStatementCacheQueries=0
