+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/bootstrap
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+./store-spring-graal $_HANDLER
