+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/java/com/unicorn/store/data/UnicornRepository.java
@@ -1,0 +1,9 @@
+package com.unicorn.store.data;
+
+import com.unicorn.store.model.Unicorn;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UnicornRepository extends CrudRepository<Unicorn, String > {
+}
