+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/java/com/unicorn/store/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,5 @@
+package com.unicorn.store.exceptions;
+
+public class ResourceNotFoundException extends RuntimeException{
+
+}
