+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/java/com/unicorn/store/model/Unicorn.java
@@ -1,0 +1,55 @@
+package com.unicorn.store.model;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Table;
+
+@Table(value = "unicorns")
+public class Unicorn {
+
+    @Id
+    private String id;
+    private String name;
+    private String age;
+    private String size;
+    private String type;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAge() {
+        return age;
+    }
+
+    public void setAge(String age) {
+        this.age = age;
+    }
+
+    public String getSize() {
+        return size;
+    }
+
+    public void setSize(String size) {
+        this.size = size;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id =  id;
+    }
+}
