+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/java/com/unicorn/store/model/UnicornEventType.java
@@ -1,0 +1,5 @@
+package com.unicorn.store.model;
+
+public enum UnicornEventType {
+    UNICORN_CREATED, UNICORN_UPDATED, UNICORN_DELETED
+}
