+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/resources/META-INF/native-image/com.amazonaws/aws-lambda-java-runtime-interface-client/jni-config.json
@@ -1,0 +1,11 @@
+[
+  {
+    "name":"com.amazonaws.services.lambda.runtime.api.client.runtimeapi.LambdaRuntimeClientException",
+    "methods":[{"name":"<init>","parameterTypes":["java.lang.String","int"] }]
+  },
+  {
+    "name":"com.amazonaws.services.lambda.runtime.api.client.runtimeapi.InvocationRequest",
+    "fields":[{"name":"id"}, {"name":"invokedFunctionArn"}, {"name":"deadlineTimeInMs"}, {"name":"xrayTraceId"}, {"name":"clientContext"}, {"name":"cognitoIdentity"}, {"name":"content"}],
+    "allPublicMethods":true
+  }
+]
