+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/resources/META-INF/native-image/com.amazonaws/aws-lambda-java-runtime-interface-client/native-image.properties
@@ -1,0 +1,1 @@
+Args = --initialize-at-build-time=jdk.xml.internal.SecuritySupport
