+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/resources/META-INF/native-image/com.amazonaws/aws-lambda-java-runtime-interface-client/resource-config.json
@@ -1,0 +1,19 @@
+{
+  "resources": {
+    "includes": [
+      {
+        "pattern": "\\Qaarch64/aws-lambda-runtime-interface-client.glibc.so\\E"
+      },
+      {
+        "pattern": "\\Qaarch64/aws-lambda-runtime-interface-client.musl.so\\E"
+      },
+      {
+        "pattern": "\\Qx86_64/aws-lambda-runtime-interface-client.glibc.so\\E"
+      },
+      {
+        "pattern": "\\Qx86_64/aws-lambda-runtime-interface-client.musl.so\\E"
+      }
+    ]
+  },
+  "bundles": []
+}
