+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/resources/META-INF/native-image/software.amazon.awssdk/aws-crt-client/resource-config.json
@@ -1,0 +1,10 @@
+{
+  "resources": {
+    "includes": [
+      {
+        "pattern": "\\Qlinux/x86_64/libaws-crt-jni.so\\E"
+      }
+    ]
+  },
+  "bundles": []
+}
