+++ labs/unicorn-store/software/alternatives/unicorn-store-spring-native-graalvm/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.datasource.url=local
+spring.datasource.username=postgres
+spring.datasource.password=test
+server.error.include-message=always
+spring.jpa.open-in-view=false
+spring.datasource.hikari.maximumPoolSize=1
+spring.datasource.hikari.data-source-properties.preparedStatementCacheQueries=0
