+++ labs/unicorn-store/software/unicorn-store-spring/pom.xml
@@ -28,7 +28,7 @@
         <dependency>
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>eventbridge</artifactId>
-            <version>2.17.281</version>
+            <version>2.17.282</version>
             <exclusions>
                 <exclusion>
                     <groupId>software.amazon.awssdk</groupId>
@@ -43,7 +43,7 @@
         <dependency>
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>aws-crt-client</artifactId>
-            <version>2.17.261-PREVIEW</version>
+            <version>2.17.282-PREVIEW</version>
         </dependency>
 
         <dependency>
