+++ labs/unicorn-store/test-app.sh
@@ -45,4 +45,19 @@ then
 fi
 
 
+if [ $app == "spring-native-graalvm" ]
+then
+    curl --location --request POST $(cat infrastructure/cdk/target/output.json | jq -r '.UnicornStoreSpringApp.ApiEndpointSpringNative')'/unicorns' \
+  --header 'Content-Type: application/json' \
+  --data-raw '{
+    "name": "Something",
+    "age": "Older",
+    "type": "Animal",
+    "size": "Very big"
+}' | jq
+
+  exit 0
+fi
+
+
